+++ client/.env
@@ -1,0 +1,3 @@
+REACT_APP_KAKAO_API_KEY= dcd41c1d660a90d5d532b21feb1e8635
+REACT_APP_KAKAO_CLIENT_SECRET= QIU5utrv9JLsejKIncCHFtL93LeCYx13
+REACT_APP_KAKAO_REDIRECT_URL= http://localhost:3000/login/oauth2/code/kakao
