+++ client/src/api/getTestUser.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import axiosInstance from "./axios_config";
+import axiosInstance from "./axios-config";
 
 export const getTestUser = async () => {
   const response = await axiosInstance({
