+++ client/src/api/time-record-api.js
@@ -1,0 +1,8 @@
+import axiosInstance from "./axios-config";
+
+export const saveTime = async (timeInfo) => {
+  const response = await axiosInstance.post("/time-record", {
+    timeInfo,
+  });
+  return response;
+};
