+++ client/src/pages/MainHome/MainHome.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 import { useSetRecoilState } from "recoil";
 import { bookmarkState } from "@recoil/bookmark-state";
-import { getBookmark } from "@api/study-room-api";
+import { getBookmark } from "@api/bookmark-api";
 import { getTestUser } from "@api/getTestUser";
 import { Bookmark, Footer, Header, StudyRoomList } from "@components/home";
-import styles from "./main_home.module.css";
+import styles from "./MainHome.module.css";
 
 function MainHome(props) {
   const setBookmark = useSetRecoilState(bookmarkState);
